ml_pipeline/requirements.txt
torch
transformers
Pillow
pinecone
tqdm
